requirements.txt
chart_studio==1.1.0
cufflinks==0.17.3
matplotlib==3.6.3
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
plotly==5.12.0
requests==2.28.2
scikit_learn==1.2.0
seaborn==0.12.2
streamlit==1.17.0
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
